Fix smallblurb wrapping around preceding lists
When a smallblurb paragraph follows a list, on Epiphany (the standard browser on the Raspberry Pi), the smallblurb wraps on the right hand side of the list. This wrapping is only a few characters wide and extremely ugly. Putting "clear: both;" into smallblurb forces the smallblurb paragraph out of the margin of the list.