Modify show_module_includes() to move modcount
modcount++ has been moved to only increment when the included
file is encountered. Separate commit in case someone disagrees.
Because of how the list of modules is generated, this change should
not affect rachel-index.php includes, but without it, modcount would
show the number of modules with rachel-index not with the inlusion file.
Alternatives would be to track another variable and/or rename modcount.