Add method for system wide and per file debug
Set debug variable in included file admin/debug.php
show single file override option in index.php
global $debug in function calls
wrap error_log commands with if($debug)
demonstrate concise error_log output with function name, var and json

This is a proposal for one way to do this.
This change is being committed last to give the option of not including
it, but is reflective of how our team (PG) would normally handle debugging
where the main advantage is leaving debug lines in the code thus
motivating people to make them descriptive and making it easy to turn
them on or off.
If there are betters ways to do this, we'll adapt to your preference.